MangoApp: Send the current upscaler instead of FSR
Changed types of used enums and fsr sharpness to uint8_t, to prevent future mistakes

+++ src/main.cpp
@@ -249,7 +249,7 @@ GamescopeUpscaleScaler g_upscaleScaler = GamescopeUpscaleScaler::AUTO;
 
 GamescopeUpscaleFilter g_wantedUpscaleFilter = GamescopeUpscaleFilter::LINEAR;
 GamescopeUpscaleScaler g_wantedUpscaleScaler = GamescopeUpscaleScaler::AUTO;
-int g_upscaleFilterSharpness = 2;
+uint8_t g_upscaleFilterSharpness = 2;
 
 bool g_bBorderlessOutputWindow = false;
